chore: move all k6 scripts
