+++ src/performanceTest/dummy-data/src/schema.js
@@ -30,7 +30,7 @@ const event = new Table('event', [
 const user = new Table('user', [
     new Field('id', (i) => `${i + 1}`),
     new Field('email', (i) => `K6-${i + 1}@email.com`),
-    new Field('name', (i) => `${i + 1}`),
+    new Field('name', (i) => `${i + 1}-name`),
     new Field('phone_number', (i) => "010-1234-1234"),
     new Field('pw', (i) => userpassword[i % 1000]),
     new Field('created_at', (i) => faker.date.recent()),
