+++ src/performanceTest/scripts/loadTest/loadTest.js
@@ -1,9 +1,8 @@
 import { check } from "k6";
-import Request from "./lib/request.js";
-import { encode } from "./lib/jwt.js";
-import hooks from "./lib/hooks.js";
-import { isSuccess, getOneFromList, randomInt } from "./lib/helpers.js";
-import exec from 'k6/execution';
+import Request from "../lib/request.js";
+import generator from "../lib/generator.js";
+import hooks from "../lib/hooks.js";
+import { isSuccess, getOneFromList } from "../lib/helpers.js";
 
 export const setup = hooks.setup
 export const handleSummary = hooks.handleSummary
@@ -21,18 +20,14 @@ export const options = {
       ]
     }
   },
-  scenarios: {
-    contacts: {
-      executor: 'per-vu-iterations',
-      vus: 200,
-      iterations: 1,
-      maxDuration: '1m', 
-    },
-  },
+  stages: [
+    { duration: '3m', target: 200 }, // traffic ramp-up from 1 to ${n} users over ${n} minutes.
+    { duration: '30s', target: 0 }, // ramp-down to 0 users
+  ],
   
   thresholds: {
     http_req_failed: ['rate<0.01'], // http errors should be less than 1%
-    http_req_duration: ['p(95)<300'], // 95% of requests should be below 300ms
+    http_req_duration: ['p(95)<200'], // 95% of requests should be below 100ms
   },
 };
 
@@ -40,16 +35,14 @@ export default function () {
   const req = new Request()
 
   const getAvaliableReservation = () => {
-    let count = 0
-    while (count < 10) {
-      req.getEvents()
-      count++
-    }
     const events = req.getEvents()
     return getOneFromList(events.json())
   }
 
-  req.setToken(encode(exec.vu.idInTest))
+  const user = generator.User()
+  req.signup(user)
+  req.signin(user)
+  check(req.access_token_info(), {"Success SignIn": isSuccess});
 
   const event = getAvaliableReservation()
   if (event) {
