+++ src/performanceTest/scripts/smokeTest/healthCheck.js
