+++ src/performanceTest/scripts/smokeTest/jwtCheck.js
